requirements.txt
numpy>=1.24
scipy>=1.10
PyYAML>=6.0
imageio>=2.31
OpenEXR>=3.3
openpyxl>=3.1
